Split base classes into their own files
