+++ lib/error/index.js
@@ -1,0 +1,15 @@
+'use strict'
+
+const ConnectionError = require('./connection-error')
+const MSSQLError = require('./mssql-error')
+const PreparedStatementError = require('./prepared-statement-error')
+const RequestError = require('./request-error')
+const TransactionError = require('./transaction-error')
+
+module.exports = {
+  ConnectionError,
+  MSSQLError,
+  PreparedStatementError,
+  RequestError,
+  TransactionError
+}
